# packages/shared/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResendConfirmationRequest": {
        "description": "Body for `POST /auth/resend-confirmation`.\n\nReplaces the previous bare `dict[str, str]` so the email is validated\nand the request schema is precise in the generated client.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "ResendConfirmationRequest",
        "type": "object"
      },
      "StatusResponse": {
        "description": "Generic `{\"status\": \"...\"}` envelope.\n\nReused by the auth endpoints (login / logout / confirm /\nresend-confirmation) which all report progress through a single\n`status` string (`ok`, `confirmed`, `already_confirmed`,\n`sent_if_unconfirmed`).",
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Bullet Digital Media onboarding orchestration API. Phase 1 - sales call to campaign go-live automation.",
    "title": "bullet-api",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/confirm/{token}": {
      "post": {
        "description": "Flip `email_confirmed` and stamp `email_confirmed_at` for the user\nencoded in `token`. Idempotent for already-confirmed users.",
        "operationId": "confirm_email_auth_confirm__token__post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Email",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/resend-confirmation": {
      "post": {
        "description": "Re-send the confirmation email for an unconfirmed user.\n\nBody: `{\"email\": \"...\"}` (validated as an email address; a malformed\nor missing value is rejected with 422 by request validation). Always\nreturns 200 with the same status regardless of whether the email\nmatched a real user - this keeps the endpoint from being a free\nuser-enumeration oracle.",
        "operationId": "resend_confirmation_auth_resend_confirmation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resend Confirmation",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}